Conflicts:
	web/src/main/java/com/softserve/edu/controller/calibrator/CalibratorVerificationController.java
	web/src/main/webapp/resources/app/admin/controllers/MeasuringEquipmentAddModalControllerAdmin.js
	web/src/main/webapp/resources/app/admin/controllers/MeasuringEquipmentControllerAdmin.js
	web/src/main/webapp/resources/app/calibrator/services/CalibrationTestServiceCalibrator.js

+++ repository/src/main/java/com/softserve/edu/device/test/data/DeviceTestData.java
@@ -21,7 +21,7 @@ public interface DeviceTestData {
     String getFileName();
     int getIntegrationTime();
     int getTestCounter();
-    long getInstallmentNumber();
+    int getInstallmentNumber();
     String getCurrentCounterNumber();
     double getLatitude();
     double getLongitude();
